almost complete function, with testing of Car and Scheduler

+++ smart-ros-pkg/server/src/fmui/DataFromCarToServer.java
@@ -3,6 +3,6 @@ package fmui;
 public class DataFromCarToServer {
 	public int userID = -1;
 	public int taskID = 0;
-	public int latitude = 0;
-	public int longitude = 0;
+	public double latitude = 0;
+	public double longitude = 0;
 }
